@@ -11,31 +11,30 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@popperjs/core": "^2.11.5",
+    "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.2.0",
-    "bootstrap-icons": "^1.8.3",
+    "bootstrap-icons": "^1.9.1",
     "datatables.net": "^1.12.1",
     "datatables.net-bs5": "^1.12.1",
-    "highlight.js": "^11.6.0",
-    "jquery": "^3.6.0"
+    "jquery": "^3.6.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
+    "@babel/core": "^7.18.13",
+    "@babel/preset-env": "^7.18.10",
     "babel-loader": "^8.2.5",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^4.0.0",
     "file-loader": "^6.2.0",
-    "html-minimizer-webpack-plugin": "^4.0.0",
+    "html-minimizer-webpack-plugin": "^4.1.0",
     "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.4.14",
-    "postcss-loader": "^7.0.0",
-    "postcss-preset-env": "^7.7.2",
-    "sass": "^1.53.0",
+    "postcss": "^8.4.16",
+    "postcss-loader": "^7.0.1",
+    "postcss-preset-env": "^7.8.0",
+    "sass": "^1.54.5",
     "sass-loader": "^13.0.2",
-    "terser-webpack-plugin": "^5.3.3",
-    "webpack": "^5.73.0",
+    "terser-webpack-plugin": "^5.3.5",
+    "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0"
   }
 }
