@@ -57,7 +57,8 @@ def get_routes(config: NamerConfig, command_queue: Queue) -> Blueprint:
 
         res = False
         if data:
-            res = get_search_results(data['query'], data['file'], config)
+            page = data['page'] if 'page' in data else 1
+            res = get_search_results(data['query'], data['file'], config, page=page)
 
         return jsonify(res)
 
@@ -101,7 +102,10 @@ def get_routes(config: NamerConfig, command_queue: Queue) -> Blueprint:
 
         res = False
         if data:
-            res = read_failed_log_file(data['file'], config)
+            res = {
+                'file': data['file'],
+                'data': read_failed_log_file(data['file'], config)
+            }
 
         return jsonify(res)
 
