@@ -169,7 +169,7 @@ class LogModal:
         wait.until(expected_conditions.element_to_be_clickable((By.CSS_SELECTOR, 'button[class="btn btn-secondary"]')))
         self.__close = self.__driver.find_element(By.CSS_SELECTOR, 'button[data-bs-dismiss="modal"]')
         self.__cancel = self.__driver.find_element(By.CSS_SELECTOR, 'button[aria-label="Close"]')
-        self.__log_item = self.__driver.find_element(By.CSS_SELECTOR, value='pre code')
+        self.__log_item = self.__driver.find_element(By.CSS_SELECTOR, value='.modal-body')
 
     def close(self) -> 'FailedPage':
         self.__close.click()
@@ -276,7 +276,7 @@ class FailedPage:
         self.__driver = driver
         find_and_wait_until_stale(driver, By.ID, "progressBar")
         self.__refresh = wait_for_and_find(driver, By.ID, 'refreshFiles')
-        self.__noFailedFiles = find_if_present(driver, By.CSS_SELECTOR, 'div[class="col m-1 text-center"] span')
+        self.__noFailedFiles = find_if_present(driver, By.CSS_SELECTOR, '#filesResult div[class="col m-1 text-center"] span')
         if self.__noFailedFiles is None:
             self.__page_links = wait_for_and_find_all(driver, by=By.CSS_SELECTOR, value='a[class="page-link"]')
             self.__search = wait_for_and_find(driver, by=By.CSS_SELECTOR, value='input[type="search"]')
